NMEA: Rework the send loop
Instead of calling updateNMEA() at a fixed interval from the main loop,
call it on each iteration and allow the function to add messages to a
send queue that is cleared on every call. This allows sensors to be
updated at a fixed interval while e.g. AIS messages can be sent out as
soon as they are ready.