AIS: Add AIS class A reports for otherShips
This commit adds reporting of the navigational status of other ships in
the simulation over NMEA AIVDM sentences via AIS class A reports. These
reports are sent out periodically depending on the ship's speed as
required by the standard. Course changes and transmission range however
are not taken into account.

Ships are assigned a MMSI if none has been specified in the scenario.