Autopilot: Add basic Autopilot-over-NMEA functionality
This commit adds very basic autopilot logic that attempts to steer the
ship towards waypoints received over NMEA. The controller also adjusts
throttle depending on leg length and progress along the leg to reduce
overshoot and so that the ships navigates routes with a closer density
of waypoints more carefully.

The autopilot only steers using the wheel, and does not use bow
thrusters or the engines

The autopilot is behind a bc5.ini flag and only enabled if it is set to
'true'.