RadarCalculation: Make the RADAR image resolution configurable
Expose the rangeResolution variable in bc5.ini and replace the hardcoded
number of 360 scanlines per turn by an angularResolution parameter also
settable via bc5.ini.

Distinguish between the current angle (which can be a float) and the
actual scanline number, which is used to index the data array.

This commit also raises the number of maximum scans per loop from 10 to
30 to allow for greater angularResolution values to still result in a
decent number of rounds per minute of scanning (given a fast enough
cpu).