Models: Fix radar.ini radar_sensitivity values
Since the radar_sensitivity value gets parsed as an unsigned integer in
RadarCalculations.cpp, the current value of 3.0 defaults to 0, which is
then clamped to 1. This commit replaces the value in the config files
with 1, resulting in the same final value for the scanAngleStep, but
making sure that the value is parsed correctly and that it is clear from
reading existing radar.ini files that the value should be an integer.